requirements.txt
openai
python-dotenv
httpx
tavily-python
streamlit
streamlit-chat
streamlit-lottie
plotly
pandas
numpy